Updates rust-toolchain to nightly-02-20.
This includes several updates to the backend based on changes in
rustc_codegen_llvm, which brings some bug fixes and a refactor to debug
info. This fixes failing dependencies that resolved to newer versions
that failed to build on the previous nightly.

Removes depending on rustc_codegen_llvm as it can't be imported with the
new toolchain for some reason.

The example cuda kernals fail to build. Their dependencies
cause nvvm to abort with ICE.